+++ modules/planning/open_space/distance_approach_problem.h
@@ -47,7 +47,7 @@ class DistanceApproachProblem {
 
   bool Solve(const Eigen::MatrixXd& x0, const Eigen::MatrixXd& xF,
              const Eigen::MatrixXd& last_time_u, const std::size_t& horizon,
-             const float& ts, const Eigen::MatrixXd& ego,
+             const double& ts, const Eigen::MatrixXd& ego,
              const Eigen::MatrixXd& xWS, const Eigen::MatrixXd& uWS,
              const Eigen::MatrixXd& l_warm_up, const Eigen::MatrixXd& n_warm_up,
              const std::vector<double>& XYbounds,
